refactor(patina): split no_potential_component_option_typo tests into submodule
